Sauvegarde des modifications avant de continuer

+++ src/Controller/ConsultationController.php
@@ -4,7 +4,7 @@ namespace App\Controller;
 
 use App\Entity\Consultation;
 use App\Form\ConsultationType;
-use App\Repository\ConsultationRenduRepository;
+use App\Repository\ConsultationRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
@@ -14,7 +14,7 @@ use Symfony\Component\Routing\Annotation\Route;
 class ConsultationController extends AbstractController
 {
     #[Route('/consultations', name: 'consultation_index')]
-    public function index(ConsultationRenduRepository $consultationRepository): Response
+    public function index(ConsultationRepository $consultationRepository): Response
     {
         $consultations = $consultationRepository->findAll();
 
@@ -26,6 +26,8 @@ class ConsultationController extends AbstractController
     #[Route('/consultations/new', name: 'consultation_new')]
     public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_VETERINAIRE');
+
         $consultation = new Consultation();
         $form = $this->createForm(ConsultationType::class, $consultation);
         $form->handleRequest($request);
@@ -45,6 +47,8 @@ class ConsultationController extends AbstractController
     #[Route('/consultations/edit/{id}', name: 'consultation_edit')]
     public function edit(Consultation $consultation, Request $request, EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_VETERINAIRE');
+
         $form = $this->createForm(ConsultationType::class, $consultation);
         $form->handleRequest($request);
 
@@ -63,6 +67,8 @@ class ConsultationController extends AbstractController
     #[Route('/consultations/delete/{id}', name: 'consultation_delete', methods: ['POST'])]
     public function delete(Consultation $consultation, EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_VETERINAIRE');
+
         $entityManager->remove($consultation);
         $entityManager->flush();
