+++ src/Controller/HabitatController.php
@@ -5,100 +5,96 @@ namespace App\Controller;
 use App\Entity\Animal;
 use App\Entity\Habitat;
 use App\Form\HabitatType;
-use App\Repository\HabitatRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Doctrine\Persistence\ManagerRegistry;
+// use App\Service\HabitatService;
 use Symfony\Component\Security\Core\Authorization\AuthorizationCheckerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpFoundation\File\UploadedFile;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 use Symfony\Component\Routing\Annotation\Route;
-
+use App\Service\FileUploader;
 
 class HabitatController extends AbstractController
 {
-    private $doctrine;
+    private ManagerRegistry $doctrine;
+    // private HabitatService $habitatService;
+    private FileUploader $fileUploader;
+    private EntityManagerInterface $entityManager;
 
-    public function __construct(ManagerRegistry $doctrine)
+    public function __construct(ManagerRegistry $doctrine, EntityManagerInterface $entityManager, FileUploader $fileUploader)# HabitatService $habitatService
     {
         $this->doctrine = $doctrine;
+        // $this->habitatService = $habitatService;
+        $this->fileUploader = $fileUploader;
+        $this->entityManager = $entityManager;
     }
 
-    #[Route('/habitat/{id}', name: 'habitat_show', methods: ['GET'])]
+    #[Route('/habitat/show/{id}', name: 'habitat_show', methods: ['GET'])]
     public function show(Habitat $habitat, AuthorizationCheckerInterface $authChecker): Response
     {
-        // Récupérer les animaux associés à cet habitat
+        // Vérification si l'habitat existe
+        if ($habitat === null) {
+            throw $this->createNotFoundException('Habitat non trouvé');
+        }
+    
+        // Récupération des animaux associés
         $animals = $habitat->getAnimals();
-
-        // Vérifier si l'utilisateur est connecté
+    
+        // Vérification des droits d'accès
         $isUserLoggedIn = $this->getUser() !== null;
-
-         // Si l'utilisateur est connecté et a un rôle spécifique, vous pouvez afficher des options supplémentaires
-         $canEdit = $authChecker->isGranted('ROLE_ADMIN') || $authChecker->isGranted('ROLE_EMPLOYE');
-        
-
-        // Rendre la vue avec les animaux de l'habitat
-        return $this->render('habitat/show.html.twig', [
-            'habitat' => $habitat,
+        $canEdit = $authChecker->isGranted('ROLE_ADMIN') || $authChecker->isGranted('ROLE_EMPLOYE');
+    
+        return $this->render('habitat/liste.html.twig', [
+            'habitats' => [$habitat], // Changed to plural
             'animals' => $animals,
             'isUserLoggedIn' => $isUserLoggedIn,
-            'canEdit' => $canEdit,  // Autoriser certains utilisateurs à éditer
+            'canEdit' => $canEdit,
         ]);
     }
 
-
-    #[Route('/habitats', name: 'habitat_index', methods: ['GET'])]
-    public function index(HabitatRepository $habitatRepository): Response
+    #[Route('/habitat', name: 'habitat_index', methods: ['GET'])]
+    public function index(): Response
     {
-        $habitats = $habitatRepository->findAll();
+        // $habitats = $this->habitatService->getAllHabitats();
+        $habitats = $this->entityManager->getRepository(Habitat::class)->findAll();
+
         return $this->render('habitat/liste.html.twig', [
             'habitats' => $habitats,
         ]);
     }
 
-    #[Route('/new', name: 'habitat_new', methods: ['GET', 'POST'])]
+    #[Route('/habitat/new', name: 'habitat_new', methods: ['GET', 'POST'])]
     public function new(Request $request): Response
     {
-        $habitat = new Habitat();
+        $habitat = new Habitat("");
         $form = $this->createForm(HabitatType::class, $habitat);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            // Récupérer le fichier téléchargé (gestion des images)
             $files = $form->get('images')->getData();
 
             if ($files) {
                 $filenames = [];
                 foreach ($files as $file) {
                     if ($file instanceof UploadedFile) {
-                        $filename = uniqid() . '.' . $file->guessExtension();
-
-                        // Déplacer le fichier
-                        $file->move(
-                            $this->getParameter('images_directory'),
-                            $filename
-                        );
-
+                        $filename = $this->fileUploader->upload($file);
                         $filenames[] = $filename;
                     }
                 }
 
-                // Enregistrer les noms des fichiers dans l'entité
                 $habitat->setImages($filenames);
             }
 
-
-
-            // Ajouter un animal à cet habitat
-            $animalName = $request->request->get('animal_name'); // Récupérer le nom de l'animal via un paramètre POST
+            $animalName = $request->request->get('animal_name');
             if ($animalName) {
                 $animal = new Animal();
                 $animal->setName($animalName);
-                $habitat->addAnimal($animal); // Lie l'animal à l'habitat
+                $habitat->addAnimal($animal);
             }
 
-            // Persister l'entité Habitat dans la base de données
             $entityManager = $this->doctrine->getManager();
             $entityManager->persist($habitat);
             $entityManager->flush();
@@ -112,38 +108,35 @@ class HabitatController extends AbstractController
         ]);
     }
 
-    #[Route('/edit/{id}', name: 'habitat_edit', methods: ['GET', 'POST'])]
+    #[Route('/habitat/edit/{id}', name: 'habitat_edit', methods: ['GET', 'POST'])]
     public function edit(Habitat $habitat, Request $request): Response
     {
+        if ($habitat === null) {
+            throw $this->createNotFoundException('Habitat non trouvé');
+        }
+
         $form = $this->createForm(HabitatType::class, $habitat);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            // Gestion des fichiers d'image
+            // Récupération des fichiers d'images
             $files = $form->get('images')->getData();
 
             if ($files) {
                 $filenames = [];
                 foreach ($files as $file) {
                     if ($file instanceof UploadedFile) {
-                        $filename = uniqid() . '.' . $file->guessExtension();
-
-                        // Déplacer le fichier dans le dossier approprié
-                        $file->move(
-                            $this->getParameter('images_directory'),
-                            $filename
-                        );
-
+                        $filename = $this->fileUploader->upload($file);
                         $filenames[] = $filename;
                     }
                 }
 
-                // Mettre à jour les images (facultatif : supprimer les anciennes)
+                // Supprimer les anciennes images si nécessaire
                 $existingImages = $habitat->getImages();
                 foreach ($existingImages as $oldImage) {
-                    $oldImagePath = $this->getParameter('images_directory') . '/' . $oldImage;
+                    $oldImagePath = $this->getParameter('uploads_directory') . '/' . $oldImage;
                     if (file_exists($oldImagePath)) {
-                        unlink($oldImagePath); // Supprimez l'ancienne image si elle existe
+                        unlink($oldImagePath);
                     }
                 }
 
@@ -162,15 +155,20 @@ class HabitatController extends AbstractController
         ]);
     }
 
-    #[Route('/delete/{id}', name: 'habitat_delete', methods: ['POST'])]
-    public function delete(Habitat $habitat, EntityManagerInterface $entityManager): Response
+    #[Route('/habitat/delete/{id}', name: 'habitat_delete', methods: ['POST'])]
+    public function delete(Habitat $habitat, EntityManagerInterface $entityManager, Request $request): Response
     {
+        if (!$this->isGranted('ROLE_EMPLOYE') && !$this->isGranted('ROLE_ADMIN')) {
+            throw new AccessDeniedException('Accès interdit.');
+        }
         foreach ($habitat->getAnimals() as $animal) {
             $animal->setHabitat(null);
         }
-
-        $entityManager->remove($habitat);
-        $entityManager->flush();
+       
+            $entityManager->remove($habitat);
+            $this->entityManager->flush();
+            $this->addFlash('success', 'L\'habitat a été supprimé avec succès.');
+        
 
         return $this->redirectToRoute('habitat_index');
     }
